dimaarion/fk.git

+++ adminpanel/admin/classes/Files.php
@@ -15,12 +15,16 @@ class Files
     {
         $sansize = new Sansize();
         $idfiles = $sansize->getrequest('idfiles');
-        if ($idfiles) {
-            if (@unlink($idfiles)) {
-                return 'Файл ' . $idfiles . ' успешно удален.';
-            } else {
-                return 'Ошибка удаления файла.';
+        if (is_file($idfiles)) {
+            if ($idfiles) {
+                if (@unlink($idfiles)) {
+                    return 'Файл ' . $idfiles . ' успешно удален.';
+                } else {
+                    return 'Ошибка удаления файла.';
+                }
             }
+        }else{
+            
         }
     }
 }
